+++ src/graph/operators/slice.cpp
@@ -56,7 +56,7 @@ Node_handle slice(Computation_graph &rec, Node_handle A,
     std::copy(offset.begin(), offset.begin() + offset.size(),
               offset_owned.begin() + diff);
 
-    for (int i = 0; i < A_val.rank(); i++) {
+    for (size_t i = 0; i < A_val.rank(); i++) {
         if (offset_owned[i] + size_owned[i] > A_val.shape()[i]) {
             std::span<const int> size_span(size.begin(), size.size());
             std::span<const int> offset_span(offset.begin(), offset.size());
